requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv
pypdf
python-docx
unstructured